pyproject.toml
[project]
name = "claude-workflow"
version = "0.1.0"
description = "A terminal-based orchestrator for running Claude Code workflows defined in YAML files"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Michal Katanski" }]
keywords = ["claude", "workflow", "orchestrator", "automation", "tmux"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Utilities",
]
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "claude-agent-sdk>=0.1.0",
    "questionary>=2.0.0",
    "claude-code-tools>=0.1.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/michalkatanski/claude-workflow"
Repository = "https://github.com/michalkatanski/claude-workflow"
Issues = "https://github.com/michalkatanski/claude-workflow/issues"

[project.scripts]
claude-workflow = "orchestrator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]

[tool.hatch.build.targets.sdist]
include = ["orchestrator", "runner.py", "README.md", "LICENSE"]
